feat: move DetailedHikeRoute RouteUtils into their own class
